+++ src/collective/taxonomy/__init__.py
@@ -1,1 +1,3 @@
 PATH_SEPARATOR = '/'
+COUNT = u"#COUNT"
+ORDER = u"#ORDER"