+++ src/collective/taxonomy/interfaces.py
@@ -23,9 +23,18 @@ class ITaxonomy(ITranslationDomain, IVocabularyFactory):
     """Persistent local utility."""
 
     def add(language, identifier, path):
-        """
-        For example: self.add('en', 'identifier', 'What a lovely day!')
+        """Add or update entry.
 
+        For example: `self.add('en', 'identifier', 'What a lovely day!')`
+        """
+
+    def replace(language, items):
+        """Replace all items for a given language.
+
+        Items are given as (path, identifier) tuples.
+
+        Note that order will be assigned naively. It is assumed that
+        items are given in depth-first order.
         """
 
     def __setitem__(identifier, term):