+++ src/collective/taxonomy/javascripts/src/js/components/Leaf.js
@@ -4,6 +4,8 @@ import { defineMessages, injectIntl } from 'react-intl'
 import AddNodeButton from '../containers/AddNodeButton'
 import EditableValues from '../containers/EditableValues'
 import RemoveNodeButton from '../containers/RemoveNodeButton'
+import MoveDownButton from '../containers/MoveDownButton'
+import MoveUpButton from '../containers/MoveUpButton'
 
 const messages = defineMessages({
   addChildNodeLabel: {
@@ -36,6 +38,16 @@ const Leaf = ({ id, index, intl, language, parentId, title }) => (
     >
       <i className="taxonomy-icon-flow-split"></i>
     </AddNodeButton>
+    <MoveDownButton
+      id={ id }
+      index={ index }
+      parentId={ parentId }
+    />
+    <MoveUpButton
+      id={ id }
+      index={ index }
+      parentId={ parentId }
+    />
   </div>
 )